+++ src/nextjs/nextjs-wizard.ts
@@ -41,6 +41,7 @@ import {
 import { traceStep, withTelemetry } from '../telemetry';
 import { getPackageVersion, hasPackageInstalled } from '../utils/package-json';
 import { getNextJsVersionBucket } from './utils';
+import { configureCI } from '../sourcemaps/sourcemaps-wizard';
 
 export function runNextjsWizard(options: WizardOptions) {
   return withTelemetry(
@@ -279,19 +280,21 @@ export async function runNextjsWizardWithTelemetry(
     path.join(process.cwd(), 'vercel.json'),
   );
 
+  if (mightBeUsingVercel) {
+    clack.log.info(
+      "▲ It seems like you're using Vercel. We recommend using the Sentry Vercel integration to set up an auth token for Vercel deployments: https://vercel.com/integrations/sentry",
+    );
+  } else {
+    await traceStep('configure-ci', () => configureCI('nextjs', authToken));
+  }
+
   clack.outro(
     `${chalk.green('Everything is set up!')}
 
    ${chalk.cyan(
      'You can validate your setup by starting your dev environment (`next dev`) and visiting "/sentry-example-page".',
    )}
-${
-  mightBeUsingVercel
-    ? `
-   ▲ It seems like you're using Vercel. We recommend using the Sentry Vercel integration: https://vercel.com/integrations/sentry
-`
-    : ''
-}
+
    ${chalk.dim(
      'If you encounter any issues, let us know here: https://github.com/getsentry/sentry-javascript/issues',
    )}`,